requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob --pre
azure-identity --pre
azure-keyvault-secrets --pre
azure-storage-file-datalake --pre
pytz --pre
datetime --pre
numpy --pre
pandas --pre


